add warning at deleting snapshot

+++ src/aks-preview/azext_aks_preview/custom.py
@@ -3994,8 +3994,11 @@ def aks_snapshot_delete(cmd,   # pylint: disable=unused-argument
                          resource_group_name,
                          name,
                          no_wait=False):
-    logger.warning('resource_group_name: %s, snapshot_name: %s ',
-                   resource_group_name, name)
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the snapshot "{}" in resource group "{}", Are you sure?'.format(name, resource_group_name)
+    if not prompt_y_n(msg, default="n"):
+        return None
+
     return client.delete(resource_group_name, name)
 
 def aks_snapshot_list(cmd,   # pylint: disable=unused-argument
